+++ packages/hardhat-zksync-deploy/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @matterlabs/hardhat-zksync-deploy
 
+## [1.2.1](https://github.com/matter-labs/hardhat-zksync/compare/@matterlabs/hardhat-zksync-deploy-v1.2.0...@matterlabs/hardhat-zksync-deploy-v1.2.1) (2024-03-04)
+
+
+### Bug Fixes
+
+* set forceDeploy to true as a default value ([#848](https://github.com/matter-labs/hardhat-zksync/issues/848)) ([326b94f](https://github.com/matter-labs/hardhat-zksync/commit/326b94f3192ae1177f5b92c186c458df246f865e))
+* support for non-inline libraries 'export default' expressions and zksolc is not reqired for config update ([#853](https://github.com/matter-labs/hardhat-zksync/issues/853)) ([dbfe6af](https://github.com/matter-labs/hardhat-zksync/commit/dbfe6af0590d85da877a6f541244492fc0efcb67))
+* update deploy cache logic, script load path support for windows ([#865](https://github.com/matter-labs/hardhat-zksync/issues/865)) ([c30e276](https://github.com/matter-labs/hardhat-zksync/commit/c30e276903e97b12283bb742659e513bdb9dafe6))
+
 ## [1.2.0](https://github.com/matter-labs/hardhat-zksync/compare/@matterlabs/hardhat-zksync-deploy-v1.1.2...@matterlabs/hardhat-zksync-deploy-v1.2.0) (2024-02-23)